m_fix : delete useless define in builtin.h
